Support duplicate detection and Refactor the project
