+++ ext.py
@@ -1,0 +1,5 @@
+from flask_mako import MakoTemplates, render_template
+from flask_sqlalchemy import SQLAlchemy
+
+mako = MakoTemplates()
+db = SQLAlchemy()
