+++ requirements.txt
@@ -9,7 +9,9 @@ Mako==1.0.7
 MarkupSafe==1.1.0
 mysqlclient==1.3.14
 Pillow==5.4.0
+pkg-resources==0.0.0
 python-magic==0.4.15
 pytz==2018.7
+short-url==1.2.2
 SQLAlchemy==1.2.15
 Werkzeug==0.14.1
